services/movies_etl/requirements/base.txt
psycopg2-binary==2.9.1
requests==2.25.0